titan: Disable secure discard
Some users have the habit of wiping the cache every now and then.
Because of the secure discard, this operation takes quite a lot, to
the point that some users may think recovery is just stuck and force
an emergency reboot while formatting the emmc. Disable secure wipes
to speed things up.

Change-Id: I22e73edaf78ab611cd6d5bb6497602c61cf2ca82